# openapi.yaml
openapi: 3.1.0
info:
  title: crypto-gateway
  version: 0.0.0
  description: |
    Runtime-agnostic crypto payment gateway with pluggable chain adapters
    (EVM / Tron / Solana), DB adapters (D1 / libSQL / Postgres), cache
    adapters, detection strategies, and webhook delivery.

    Deploys on Cloudflare Workers, Node/Bun, Deno, and Vercel Edge from one
    codebase.

    ### Deployment knobs (not HTTP routes)

    The behaviors below are server-side config read at boot and do not
    surface as endpoints. They're documented here so operators who only
    have the OpenAPI spec can discover them.

    **Price oracle fallback chain.** Invoice quoting walks a chained list
    of providers on every request; each link tries itself first and
    delegates to the next on error/timeout/unmapped symbol. Only the
    terminal `static-peg` link serves hardcoded values. Default ordering
    when `ALCHEMY_API_KEY` is set:
    `CoinGecko → Alchemy → CoinCap → Binance → static-peg`.
    Knobs: `PRICE_ADAPTER=alchemy|coingecko|static-peg` (swaps the primary
    or collapses to the static link), `COINGECKO_API_KEY` +
    `COINGECKO_PLAN=demo|pro`, `COINCAP_API_KEY`, and per-provider
    opt-outs `DISABLE_COINGECKO=1` / `DISABLE_COINCAP=1` /
    `DISABLE_BINANCE=1` / `DISABLE_ALCHEMY=1`. `PRICE_ADAPTER=alchemy`
    without a key logs a warning and falls back to the default ordering.

    **Log shipping + alerting.** Setting `ALERT_WEBHOOK_URL` attaches an
...

...
servers:
...
security:
  - MerchantApiKey: []
paths:
...
  /webhooks/alchemy:
    post:
      tags: [webhooks]
      summary: Ingest Alchemy Notify ADDRESS_ACTIVITY webhooks
      description: |
        HMAC-SHA256 of the raw body under the per-webhook signing key
        (looked up by `event.webhookId` in the encrypted-at-rest
        `alchemy_webhook_registry` table) must match `X-Alchemy-Signature`.
        Returns 200 fast and processes the payload in the background;
        Alchemy retries on non-2xx.
      security:
        - AlchemySignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Acknowledged; ingest happens in background
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean, const: true }
        "400": { description: Malformed JSON after a valid signature }
        "401": { description: Missing or bad signature }
        "404": { description: Alchemy ingest not enabled on this deployment }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /admin/alchemy-webhooks/signing-keys:
    post:
      tags: [admin]
      summary: Register an Alchemy webhook's signing key manually
      description: |
        For operators who created a webhook through the Alchemy dashboard UI
        (not via bootstrap) or who need to push a key back in after a
        persistence failure. The signing key is encrypted via
        `SECRETS_ENCRYPTION_KEY` before being written to
        `alchemy_webhook_registry`. Re-running with the same chainId upserts
        (rotation after a dashboard delete+recreate).
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, webhookId, signingKey, webhookUrl]
              properties:
                chainId: { type: integer }
                webhookId: { type: string, minLength: 1, maxLength: 128 }
                signingKey: { type: string, minLength: 1, maxLength: 256 }
                webhookUrl: { type: string, format: uri }
      responses:
        "201":
          description: Signing key persisted encrypted-at-rest
          content:
            application/json:
              schema:
                type: object
                properties:
                  registered:
                    type: object
                    properties:
                      chainId: { type: integer }
                      webhookId: { type: string }
                      webhookUrl: { type: string, format: uri }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Admin surface not enabled (ADMIN_KEY not set) }
...
components:
  securitySchemes:
    MerchantApiKey:
      type: http
      scheme: bearer
      description: |
        Merchant API key returned once at `POST /admin/merchants`.
        Also accepted as `X-API-Key: <key>`.
    AdminKey:
      type: http
      scheme: bearer
      description: Shared admin key from `ADMIN_KEY` env. Also accepted as `X-Admin-Key`.
    AlchemySignature:
      type: apiKey
      in: header
      name: X-Alchemy-Signature
      description: HMAC-SHA256 hex of the raw body under the per-webhook signing key stored in `alchemy_webhook_registry` (looked up by `event.webhookId`).
    CronSecret:
      type: http
      scheme: bearer
      description: Shared secret from `CRON_SECRET` env.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests in the current window
      headers:
        Retry-After:
          description: Seconds until the window rolls
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code: { type: string, description: Stable machine-readable code }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Optional per-code payload (e.g. Zod issues on VALIDATION)
...